@@ -114,7 +114,7 @@
     "@types/jest": "^27.0.1",
     "@types/jsdom": "^16.2.13",
     "@types/jsonwebtoken": "^9.0.2",
-    "@types/node": "18.11.9",
+    "@types/node": "20",
     "@types/ramda": "^0.30.2",
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.37.0",
@@ -148,8 +148,8 @@
     "nodemon": "^2.0.12",
     "supertest": "^6.0.1",
     "ts-node": "^10.4.0",
-    "typescript": "^5.6.3",
-    "typescript-eslint": "^8.14.0"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0"
   },
   "standard": {
     "ignore": [
