@@ -9,8 +9,8 @@
     "clean": "rm -rf ./dist",
     "cleanbuild": "yarn clean && yarn build",
     "coverage": "yarn test:coverage",
-    "lint:fix": "yarn eslint . --ext .js,.jsx,.ts,.tsx --fix",
-    "lint": "yarn eslint . --ext .js,.jsx,.ts,.tsx",
+    "lint:fix": "yarn eslint . --fix",
+    "lint": "yarn eslint . ",
     "rebuild": "yarn cleanbuild",
     "test:coverage": "yarn jest --coverage",
     "test:integration": "yarn jest --config ./src/tests-integration/jest.integration.config.js",
@@ -27,6 +27,7 @@
   },
   "description": "Root MTC Typescript project.  Contains all shared components and azure function implementations",
   "devDependencies": {
+    "@eslint/js": "^9.16.0",
     "@faker-js/faker": "^8.4.1",
     "@types/adm-zip": "^0.5.0",
     "@types/async": "^3.2.7",
@@ -39,7 +40,7 @@
     "@types/jsonwebtoken": "^9.0.2",
     "@types/morgan": "^1.9.4",
     "@types/mssql": "^9.1.5",
-    "@types/node": "^18.11.9",
+    "@types/node": "20",
     "@types/ramda": "^0.30.2",
     "@types/ua-parser-js": "^0.7.36",
     "@types/unzipper": "^0.10.4",
@@ -47,18 +48,19 @@
     "@typescript-eslint/eslint-plugin": "^5.61.0",
     "@typescript-eslint/parser": "^5.61.0",
     "bcryptjs": "^2.4.3",
-    "eslint": "^8.44.0",
+    "eslint": "9.14.0",
     "eslint-config-standard-with-typescript": "^36.0.0",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jest": "^27.2.2",
-    "eslint-plugin-n": "^16.0.1",
-    "eslint-plugin-node": "^11",
-    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.9.0",
+    "eslint-plugin-n": "^17.14.0",
+    "eslint-plugin-promise": "^7.2.1",
+    "globals": "^15.13.0",
     "he": "^1.2.0",
     "jest": "^29.6.1",
     "node-mocks-http": "^1.12.2",
     "ts-jest": "^29.1.2",
-    "typescript": "^5.1.6"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0"
   },
   "dependencies": {
     "@azure/core-asynciterator-polyfill": "^1.0.2",
