@@ -1,6 +1,6 @@
 import config from '../config'
 import { isNil } from 'ramda'
-const appInsights = require('applicationinsights')
+import * as appInsights from 'applicationinsights'
 
 const cloudRoleName = 'TsLib'
 
@@ -29,14 +29,14 @@ if (isNil(connectionString)) {
 }
 
 const appInsightsHelper = {
-  startInsightsIfConfigured: async (cloudRole: string = 'TsLib (default)'): Promise<void> => {
+  startInsightsIfConfigured: async (cloudRole = 'TsLib (default)'): Promise<void> => {
     appInsights.defaultClient.context.tags[appInsights.defaultClient.context.keys.cloudRole] = cloudRole
     appInsights.start()
     console.log(`TSLib Application insights: started for role [${cloudRole}]`)
     let buildNumber
     try {
       buildNumber = 'NOT IMPLEMENTED'
-    } catch (error) {
+    } catch {
       buildNumber = 'NOT FOUND'
     }
     appInsights.defaultClient.commonProperties = {
