@@ -146,8 +146,7 @@ async function completeMessages (messageBatch: sb.ServiceBusReceivedMessage[], r
   // the sql updates are idempotent and as such replaying a message
   // will not have an adverse effect.
   context.log(`${logPrefix}: batch processed successfully, completing all messages in batch`)
-  for (let index = 0; index < messageBatch.length; index++) {
-    const msg = messageBatch[index]
+  for (const msg of messageBatch) {
     try {
       await receiver.completeMessage(msg)
     } catch (error) {
@@ -167,8 +166,7 @@ async function completeMessages (messageBatch: sb.ServiceBusReceivedMessage[], r
 }
 
 async function abandonMessages (messageBatch: sb.ServiceBusReceivedMessage[], receiver: sb.ServiceBusReceiver, context: InvocationContext): Promise<void> {
-  for (let index = 0; index < messageBatch.length; index++) {
-    const msg = messageBatch[index]
+  for (const msg of messageBatch) {
     try {
       await receiver.abandonMessage(msg)
     } catch (error) {
@@ -188,7 +186,7 @@ async function process (context: InvocationContext, messageBatch: sb.ServiceBusR
     const linesOfData = csvTransformer.transform()
     await psReportStagingDataService.appendDataToBlob(linesOfData)
     await completeMessages(messageBatch, receiver, context)
-  } catch (error) {
+  } catch {
     // sql transaction failed, abandon...
     await abandonMessages(messageBatch, receiver, context)
   }
