@@ -2,8 +2,8 @@ import * as path from 'path'
 import * as fs from 'fs'
 
 export class PingService {
-  private commitId: string = ''
-  private buildNumber: string = ''
+  private commitId = ''
+  private buildNumber = ''
 
   public async getCommitId (): Promise<string> {
     if (this.commitId === '') {
