@@ -55,7 +55,7 @@ describe('async-retry', () => {
       }
       try {
         await retry<number>(func, strategy, () => true)
-      } catch (error) {
+      } catch {
         expect(callCount).toBe(maxAttempts)
       }
     })
@@ -73,7 +73,7 @@ describe('async-retry', () => {
       try {
         await retry<number>(func, retryPolicy, () => true)
         expect(callCount).toBe(3)
-      } catch (error) {
+      } catch {
         fail(`should have completed after 3 attempts. attempts made:${callCount}`)
       }
     })
